backend/requirements.txt
flask
openai[aiohttp]>=1.50.0
flask_cors
mongoengine
flask_mongoengine
python-dotenv
Werkzeug==2.2.2
gunicorn
firebase_admin
PyJWT
Flask-SocketIO
flask_jwt_extended
eventlet
pymongo